backend/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Lambda functions use boto3 which is included in the AWS Lambda runtime
# No additional dependencies needed for basic functionality
